schema: Fold the remaining NamedObject code into Object
Since all schema objects are now named, it makes sense to merge the
delta ops as well.  This also stops the abuse of `AlterDatabase` as the
root delta command by adding an explicit class for that role.
